+++ projects/framework-harmony/extensions/plugins/harmony_thumbnail_exporter.py
@@ -10,22 +10,22 @@ from ftrack_framework_core.exceptions.plugin import PluginExecutionError
 
 
 class HarmonyThumbnailExporterPlugin(BasePlugin):
-    name = 'harmony_thumbnail_exporter'
+    name = "harmony_thumbnail_exporter"
 
-    SUPPORTED_FORMATS = ['bmp', 'gif', 'jpeg', 'jpg', 'png', 'tif', 'tiff']
+    SUPPORTED_FORMATS = ["bmp", "gif", "jpeg", "jpg", "png", "tif", "tiff"]
 
     def run(self, store):
-        '''
+        """
         Pick up previously rendered image sequence and pick the middle image
         for use as thumbnail if possible.
-        '''
-        component_name = self.options.get('component')
+        """
+        component_name = self.options.get("component")
         sequence_expression = (
-            store['components'].get('sequence', {}).get('exported_path')
+            store["components"].get("sequence", {}).get("exported_path")
         )
         if not sequence_expression:
             raise PluginExecutionError(
-                f'Could not locate previously exported image sequence component and output'
+                "Could not locate previously exported image sequence component and output"
             )
 
         collection = clique.parse(sequence_expression)
@@ -35,10 +35,10 @@ class HarmonyThumbnailExporterPlugin(BasePlugin):
             )
 
         # Check if the format is supported
-        extension = collection.format('{tail}')
+        extension = collection.format("{tail}")
         if extension not in self.SUPPORTED_FORMATS:
             self.logger.warning(
-                f'Unsupported thumbnail format: {extension}, skipping'
+                f"Unsupported thumbnail format: {extension}, skipping"
             )
             return
 
@@ -50,9 +50,9 @@ class HarmonyThumbnailExporterPlugin(BasePlugin):
         image_path = get_temp_path(filename_extension=extension)
 
         # Copy the frame to the temp path
-        src = collection.format('{head}%s{tail}' % middle_frame)
+        src = collection.format("{head}%s{tail}" % middle_frame)
 
-        self.logger.debug(f'Copying {src} to {image_path}')
+        self.logger.debug(f"Copying {src} to {image_path}")
         shutil.copy(src, image_path)
 
-        store['components'][component_name]['exported_path'] = image_path
+        store["components"][component_name]["exported_path"] = image_path